Adds mosquitto installation
Mosquitto deployment runs a single pod, but uses a NodePort service to expose mosquitto on all cluster nodes, thus achieving higher uptime, by setting the DNS A records to point to all the cluster node IP addresses.

This way, connecting to `mosquitto.k3s.pragmatic-software.gr` port 31883 will always take the traffic to the correct pod, regardless of the node it's running on.