Fixes HTTPS access to the unifi controller
Without this in place there was an error by the unifi application that "This combination of host and port requires TLS".

What was needed was essentially TLS passthrough from the Ingress (controller) all the way to the pod.

Traefik is "clever" like that and automatically enables TLS passthrough if it finds that the backend service name starts with "https" !!!

From the docs:

> If the service port defined in the ingress spec has a name that starts with https (such as https-api, https-web or just https).

https://doc.traefik.io/traefik/routing/providers/kubernetes-ingress/#communication-between-traefik-and-pods