enhances traefik deployment
through HelmChartConfig.

Switches traefik deployment to DaemonSet (for HA) and also persists the `--serversTransport.insecureSkipVerify=true` option that is necessary in order to expose non-https internal services.

This also exposes the Traefik dashboard for help with debugging (with no auth middleware for now, since this is only an internal network) and also makes a first attempt at a UDP ingress.